feat: provide `network-id` input
This is required for the bootstrap network to be able to connect to the original.

If we want to connect to the `mainnet`, we can just explicitly use `1` as the value. This is better
than making it optional, because we will inevitably forget to provide it, and most of the time we
won't be bootstrapping from `mainnet`.