chore: provide `funding-wallet-secret-key` input
There are two different inputs related to the funding wallets now. You can either supply the secret
of a wallet that will fund other wallets that get created during the deploy process, or you can
supply secret keys for pre-funded wallets. The latter is useful for working with the production
network, just so you can have more control over the management of the funds.