chore: check the version returned by crates.io
It seems the case that the request to retrieve the latest version from crates.io possibly gets rate
limited when the request is made from Github. If I run the same request from my dev machine, it is
fine, but when run from Github it is returning the string "null".

If the version hasn't been retrieved correctly, we can just use the latest code rather than check
out a specific version.